mapstruct+ price int to big decimal
Changed primary keys of entity from int to jpa standard Long and price type of int to industry standard bigdecimal also in database to decimal(10,2) in flyway V3

add mappers using mapstruct for entities and dtos
and tested them
added mysql testing container

+++ src/main/java/org/iti/ecomus/dto/OrderDTO.java
@@ -5,6 +5,8 @@ import lombok.*;
 import org.iti.ecomus.entity.OrderDetails;
 import org.iti.ecomus.enums.OrderStatus;
 import org.iti.ecomus.enums.PayType;
+
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -13,11 +15,11 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDTO {
-    private int orderId;
+    private Long orderId;
 
     private String address;
 
-    private int price;
+    private BigDecimal price;
 
     private Date date;
